+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Expose gRPC metrics (#584)
+- Expose database transaction and events metrics (#589)
 
 ## [0.6.1] - 2022-03-01